I dont want the Card Preview to be contained on the customers phone when they scan the QR code the card must fit neatly on their phone without any borders or containers.
So what we can do is scale the design to fit the resolution of the device that the card is being viewed on.

I also want to be able to click on the share button at the bottom right and the QR code must pop up with the ability to X out of it when im done.

I also want the UI to be very similar in colour to the splash screen