can you make the suggest designs functional.
I want to take all the information that people have uploaded.Gemini must search all of the links and change the card preview.

That means It must suggest background,about me section and anything else necessary to have on the card.

i also would prefer to have all the Icons of the potential info the client would have on their card and only if the client clicks on that icon and enters their details in the backend,will it show up on the card preview.

That way the card preview can be correctly spaced and adjusted to accomodate the actual platforms that the client uses.