Okay so I want the admin to send them this invite link.
The staff member will open it and verify their email address.then they will use the access code.

This will send them to a limited detail fill in sheet.

So this will get all their information necessary to have on the digital business card.

This information will later be used in the design so we want as much information as possible to ensure we are up to date with the staff members details.

They will also be able to ask the admin to send them a edit link where they would be able to update their details if anything had to change.

Once these details are collected and saved it would automatically sync with their design to ensure that the most up to date details would be present on their digital business card